Updated installer maker
Fixed issue #3.
Further Safetied the installer-maker.
Added options to run the installer maker with command-line options, falling back on defaults that allows it to source from the directory it is running in.
Removed responsibility of archiver writing and updating the installer file - its only goal should be to retrieve, merge, and prepare a buffer for all files to-be-archived (and the reverse), but nothing related to the installer files.