Added UAC control and open dir checkbox
Added UAC control so that it can be run as an administrator, in case we're installing to a protected folder like \\Program files\\

Added a checkbox, which doesn't do anything yet. Later, I'm going to probably try to implement some sort of state design pattern to manage all these buttons.