Changed header processing
- Header base class
- Compression/Diff/Package/Patch headers
- NST Compression functions now use the buffer class as parameters
- NST Diff functions now use the buffer class as parameters
- Added tie-in where cmake generates a version values, derived from PROJECT_VERSION, producing a new Resource.h file

This update was 2 days in the making. I had to deal with a lot of bugs along the way, so some things may look weird. After a long night, everything seems to be functional, but some inconsitencies may be found in the code. I'll inevitably do another code clean-up soon enough anyways.