1.6.23
Updated threader for code coverage reasons
Discovered that I don't need to detach the threads in this circumstance, joining them is good enough.
To that end, I don't need to track the number of active threads, since I signal for them to finish and then join them to the main thread.